Create the contact model class
Create a class named contact.model.ts to store all of the data associated with an individual contact. This class is analogous to the Recipe model class in the Recipe Book application.

Name the class Contact.

Inside of the Contact class, create a method named constructor. This will be the default constructor method of the class.
Inside of the parentheses of the constructor method, add the following arguments:

id—the id of the contact.

name—the name of the contact.

email—the email address of the contact.

phone—the phone number of the contact.

imageUrl—the URL of the photo image of the contact.

group—this attribute is only applicable to group contacts. It is an array of other contacts that belong to the group.

Make sure to add the access modifier public before each of the arguments defined above. This will make them public properties of the class.